Reach zero clippy warnings workspace-wide; bump MSRV to 1.82
cargo clippy --workspace --all-targets goes from ~250 warnings plus 7
denied errors to zero, with behavior parity to the C oracle preserved
throughout:

- rust-version 1.76 -> 1.82 (nothing pinned 1.76; satisfies all
  incorrect-msrv-lint warnings on the 1.97 toolchain)
- mechanical fixes where provably equivalent: contains_key/contains
  rewrites, doc list indentation, manual range contains, derivable
  Default impls, unnecessary casts, let-else to ?, or-patterns
- targeted #[allow] with justification comments (citing C source
  locations) where a fix would change or obscure ported semantics:
  PPD offset tables (identity_op/erasing_op), C early-return mirrors,
  too_many_arguments on legacy driver signatures, C-mirror test
  formulas, the intentional shrine_type >= 255 comparison
  (area/14/random.c:2570), and copy-shaped if/match chains that map
  one-to-one onto C case statements
- oracle-verified lint sites from the bug hunt (identical if blocks,
  indexed loops, manual clamps) annotated with their C file:line
  provenance

No logic changes; test suite stays at full green.

+++ crates/ugaris-core/src/world/light.rs
@@ -44,7 +44,7 @@ impl World {
 
     pub(crate) fn mark_dirty_sector(&mut self, x: usize, y: usize) {
         self.dirty_sectors
-            .set_sector(x as isize, y as isize, self.tick.0.max(1) as u64);
+            .set_sector(x as isize, y as isize, self.tick.0.max(1));
     }
 
     pub(crate) fn mark_light_area(&mut self, x: usize, y: usize, strength: i16) {
